pyproject.toml
[project]
name = "hackaton"
version = "0.1.0"
description = ""
authors = [
    {name = "rafateste-web",email = "rafael.o.sebastiao@aluno.senai.br"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "tzdata (>=2025.3,<2026.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = "fastapi dev hackaton/app.py"
makemigration = "alembic revision --autogenerate -m"
upgrade = "alembic upgrade +1"
downgrade = "alembic downgrade -1"
create_admin = "python -u create_admin.py"